Benchmark 2 implemented
I attempted to implement benchmark 2, but I am not confident in my mask for the initial condition - I think that I should whittle it down a little to ensure there's only one node in thickness all around the circle. Otherwise, I think it is fairly close and looks approximately similar.