Finished benchmark 3 and progress with Inamuro b.c.
For benchmark 3, I increased niter to 50000 which did the job and I manually resized the plot instead of trying to program it. I modified benchmark2_plot to allow for the inamuro b.c. but I haven't managed to get the simulation of it to work: I think I finally understand how it's meant to work but I haven't thought of a nice way to program it yet without a huge amount of if statements for each case around the mask.