work on heat benchmark
i got heat benchmark to run without temperature heading to infinity. a homogenous temperature stays as it is, but fluctuations lead to a slight decrease over time of the temperature overall - probably due to the boundary conditions or the collisions within the liquid due to forcing. yet to compare to benchmark. was helped on the way by creating a linear version of every equation in 'stripped' which doesn't include linearities - this is just for debugging.