Working boundary conditions
Made Inamuro B.c. work for the curve of the finger. To do this I had to increase ny to 66 to have a spare node at the top and bottom to be able to find the directions inwards.
Main parts of Inamuro code:
1) creating mask for the outline of the finger
2) find out the directions inwards at each node in the mask around the curved surface at the end of the finger
3) implementing an algorithm in the boundary condition part of code to carry it out
I also added top and bottom boundary conditions to the finger to ensure cooling from all directions.
In plotting, I changed the way the outline of the finger was plotted from being a contour to being a few lines, not sure if I need it at all to be honest. I also changed the plotting of the temperature from a contour plot to a nicer colour coded plot using inbuilt functions for temperature.