Tidy and necrosis
I tidied up the boundary conditions for the zou and he bcs at the edge of the finger since having skin move is pretty unphysical so it's easier to read.
I tried to add a necrosis function as in the 3D paper but it doesn't appear to work, unsure why.